Adding force refresh of frontpage weather upon selection of default location
