+++ src/contexts/UserDataContext.jsx
@@ -1,20 +1,37 @@
-import React, { useContext, useMemo, useEffect } from 'react';
+import React, {
+    useContext, useMemo, useEffect, useState,
+} from 'react';
 import PropTypes from 'prop-types';
+import axios from 'axios';
 
 export const UserDataContext = React.createContext();
 
-export function UserDataProvider({
-    children, fetchUserData, userData, setUserData,
-}) {
+export function UserDataProvider({ children }) {
+    const [userData, setUserData] = useState(null);
+
+    const fetchUserData = async () => {
+        const backendUrl = process.env.REACT_APP_API_URL || 'http://localhost:3000';
+        try {
+            const res = await axios.get(`${backendUrl}/user/dashboard`, {
+                headers: {
+                    'x-auth-token': localStorage.getItem('token'),
+                },
+            });
+            setUserData(res.data.user);
+        } catch (err) {
+            console.error(err.response.data);
+        }
+    };
+
+    useEffect(() => {
+        fetchUserData();
+    }, []);
+
     const contextValue = useMemo(() => ({
         userData,
         setUserData,
         fetchUserData,
-    }), [userData, setUserData, fetchUserData]);
-
-    useEffect(() => {
-        console.log('UserDataContext updated:', userData);
-    }, [userData]);
+    }), [userData]);
 
     return (
         <UserDataContext.Provider value={contextValue}>
@@ -25,22 +42,6 @@ export function UserDataProvider({
 
 UserDataProvider.propTypes = {
     children: PropTypes.node.isRequired,
-    fetchUserData: PropTypes.func.isRequired,
-    userData: PropTypes.shape({
-        id: PropTypes.string,
-        username: PropTypes.string,
-        email: PropTypes.string,
-        settings: PropTypes.shape({
-            defaultCountry: PropTypes.string,
-            defaultLocation: PropTypes.string,
-            temperatureUnit: PropTypes.string,
-        }),
-    }),
-    setUserData: PropTypes.func.isRequired,
-};
-
-UserDataProvider.defaultProps = {
-    userData: null,
 };
 
 export const useUserData = () => {
